chore(api): remove outdated CMS documentation and project management mutations
